requirements.txt
# CarePill 약품 관리 시스템 의존성

# 기존 음성 채팅 시스템
openai>=1.0.0
pyaudio>=0.2.11
pygame>=2.5.0

# 컴퓨터 비전 및 이미지 처리
opencv-python>=4.8.0
pytesseract>=0.3.10
ultralytics>=8.0.0  # YOLO
Pillow>=10.0.0
numpy>=1.24.0

# 데이터베이스
SQLAlchemy>=2.0.0
sqlite3  # Python 내장
mysql-connector-python>=8.0.0
alembic>=1.12.0  # 데이터베이스 마이그레이션

# 웹 프레임워크 (Django)
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# 데이터 처리
pandas>=2.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# 하드웨어 제어 (라즈베리파이)
RPi.GPIO>=0.7.1
gpiozero>=1.6.2
picamera2>=0.3.15

# 시스템 및 유틸리티
python-dotenv>=1.0.0
schedule>=1.2.0
logging>=0.4.9.6
configparser>=6.0.0
pydantic>=2.0.0

# 테스팅
pytest>=7.4.0
pytest-django>=4.5.0
pytest-cov>=4.1.0
mock>=5.1.0

# 개발 도구
black>=23.7.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.5.0

# 배포 및 모니터링
gunicorn>=21.2.0
psutil>=5.9.0
sentry-sdk>=1.32.0

# API 문서화
drf-spectacular>=0.26.0